fixed bug in logger - now determining leaf nodes in AST abbreviation and assigning them a loop depth value - prolly not right at the moment

+++ include/Logger.hpp
@@ -17,18 +17,18 @@ namespace coop{
         extern std::ostream& out_stream;
         extern std::stringstream log_stream;
         //configurable depth (out message indentation)
-        extern size_t depth;
+        extern int depth;
 
         void clear(std::stringstream& msg_stream = log_stream);
 
         /*outs message to out_stream*/
-        size_t& out(const char* msg, const char* append = "\n");
+        int& out(const char* msg, const char* append = "\n");
 
         /*outs message to out_stream*/
-        size_t& out(std::stringstream& msg_stream, const char* append = "\n");
+        int& out(std::stringstream& msg_stream, const char* append = "\n");
 
         /*outs log_stream to out_stream*/
-        size_t& out();
+        int& out();
 
         void err(Should_Exit se = Should_Exit::YES);
 
@@ -36,10 +36,10 @@ namespace coop{
         /*informing the user of a progress*/
         enum Progress_Status {RUNNING, DONE, TODO};
 
-        size_t& out(const char* msg, Progress_Status status);
+        int& out(const char* msg, Progress_Status status);
 
         /*outs message to out_stream informing the user of a progress*/
-        size_t& out(std::stringstream& msg_stream, Progress_Status status);
+        int& out(std::stringstream& msg_stream, Progress_Status status);
 
         void out(Progress_Status status);
     }
